+++ WormGame/Pooling/PoolableEntity.cs
@@ -4,7 +4,7 @@ using Otter.Graphics;
 namespace WormGame.Pooling
 {
     /// @author Antti Harju
-    /// @version 18.07.2020
+    /// @version 30.07.2020
     /// <summary>
     /// Class for poolable entities.
     /// </summary>
@@ -34,7 +34,7 @@ namespace WormGame.Pooling
         /// <summary>
         /// Sets entity visibility and various other Otter2d entity properties.
         /// </summary>
-        public bool Enabled { get { return Visible; } set { AutoUpdate = value; AutoRender = value; Visible = value; } }
+        public virtual bool Enabled { get { return Visible; } set { AutoUpdate = value; AutoRender = value; Visible = value; } }
 
 
         /// <summary>